requirements.txt
alembic==1.11.1
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
boto3==1.26.160
botocore==1.29.160
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contourpy==1.1.0
cycler==0.11.0
databricks-cli==0.17.7
debugpy==1.6.5
decorator==5.1.1
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.1.3
entrypoints==0.4
executing==1.2.0
filelock==3.12.2
FileManager==1.0.12
Flask==2.3.2
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.19.4
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.2
kiwisolver==1.4.4
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mlflow==2.4.1
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==22.0
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.23.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==12.0.1
Pygments==2.14.0
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
querystring-parser==1.2.4
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
sh==2.0.4
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.8.0
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.121
urllib3==1.26.16
waitress==2.1.2
wcwidth==0.2.5
websocket-client==1.6.0
Werkzeug==2.3.6
zipp==3.15.0
